second beta version - improved server logic
